Refactor #62 - common패키지 생성 및 관련 파일 이동
common패키지를 생성해서 AwsS3Service클래스를 옮김
